add image for k8s-based dataflow deployment
Some dataflow jobs need to be deployed via k8s as there are significant limitations
on how Scio behaves on Dataflow (mostly optional parameter handling).
The kind of deployment has been a problem before. As the terraform resources are
created asynchronously the jobs would fail for missing temporary buckets that
are required to start. Therefore we need some sort of backoff strategy that will
delay the deployment for a couple of seconds before the actual deployment trigger happens.